+++ modules/hanna-css/src/lib/style-server-info.ts
@@ -75,6 +75,7 @@ export type CssModuleToken =
   | 'NewsHero'
   | 'PageFilter'
   | 'PageHeading'
+  | 'Pagination'
   | 'Picture'
   | 'ProcessOverview'
   | 'PullQuote'